feat(codegen): add doc comments for ts morph generator code
