+++ packages/codegen/src/ts-morph-utils/file-builders/editor-common.ts
@@ -1,17 +1,47 @@
+import { pascalCase } from "change-case";
 import path from "path";
 import type { Project } from "ts-morph";
-import { makeModulesFile } from "./module-files.js";
+import { documentEditorModuleFileTemplate } from "../templates/document-editor/module.js";
+import { getOrCreateSourceFile } from "../ts-morph-project.js";
 
-export function makeEditorsModulesFile(project: Project, projectDir: string) {
-  const modulesDirPath = path.join(projectDir, "editors");
-  const modulesSourceFilesPath = path.join(modulesDirPath, "/**/*");
-  makeModulesFile({
-    project,
-    modulesDirPath,
-    modulesSourceFilesPath,
-    outputFileName: "editors.ts",
-    typeName: "EditorModule",
-    variableName: "editors",
-    variableType: "EditorModule[]",
+type MakeEditorModuleFileArgs = {
+  project: Project;
+  editorName: string;
+  editorId: string;
+  documentModelId?: string;
+  editorDirPath: string;
+  legacyMultipleDocumentTypes?: string[];
+};
+/** Generates the `module.ts` file for a document editor or drive editor */
+export function makeEditorModuleFile({
+  project,
+  editorDirPath,
+  editorName,
+  documentModelId,
+  editorId,
+  legacyMultipleDocumentTypes,
+}: MakeEditorModuleFileArgs) {
+  if (documentModelId && !!legacyMultipleDocumentTypes) {
+    throw new Error(
+      "Cannot specify both documentModelId and legacyMultipleDocumentTypes",
+    );
+  }
+  const filePath = path.join(editorDirPath, "module.ts");
+  const { sourceFile } = getOrCreateSourceFile(project, filePath);
+
+  sourceFile.replaceWithText("");
+
+  const pascalCaseEditorName = pascalCase(editorName);
+  const documentTypes = documentModelId
+    ? `["${documentModelId}"]`
+    : JSON.stringify(legacyMultipleDocumentTypes);
+
+  const template = documentEditorModuleFileTemplate({
+    editorName,
+    editorId,
+    pascalCaseEditorName,
+    documentTypes,
   });
+  sourceFile.replaceWithText(template);
+  project.saveSync();
 }
